Account component renamed.
